+++ src/main/uni_hid_device.c
@@ -497,6 +497,8 @@ void uni_hid_device_guess_controller_type_from_pid_vid(uni_hid_device_t* d) {
       logi("Device detected as Wii controller: 0x%02x\n", type);
       break;
     case CONTROLLER_TYPE_SwitchProController:
+    case CONTROLLER_TYPE_SwitchJoyConRight:
+    case CONTROLLER_TYPE_SwitchJoyConLeft:
       d->report_parser.setup = uni_hid_parser_switch_setup;
       d->report_parser.init_report = uni_hid_parser_switch_init_report;
       d->report_parser.parse_raw = uni_hid_parser_switch_parse_raw;